requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.2
azure-storage-blob==12.20.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.2
environs==9.5.0
fastapi==0.110.3
fastapi-cli==0.0.4
fastjsonschema==2.19.1
Flask==3.0.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipython==8.12.3
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyterlab_pygments==0.3.0
langchain==0.2.3
langchain-cli==0.0.25
langchain-community==0.2.4
langchain-core==0.2.5
langchain-experimental==0.0.60
langchain-milvus==0.1.1
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langserve==0.2.2
langsmith==0.1.76
libcst==1.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
minio==7.2.7
mistune==3.0.2
MouseInfo==0.1.3
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
openai==1.33.0
orjson==3.10.4
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
pathspec==0.12.1
pillow==10.3.0
pipreqs==0.5.0
protobuf==4.25.3
pyarrow==16.1.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
PyGetWindow==0.0.9
pymilvus==2.4.4
PyMsgBox==1.0.9
pypdf==4.2.0
pyperclip==1.9.0
pyproject-toml==0.0.10
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
python-multipart==0.0.9
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.1
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sse-starlette==1.8.2
starlette==0.37.2
streamlit==1.35.0
tenacity==8.3.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenize-rt==5.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
toolz==0.12.1
tqdm==4.66.4
typer==0.9.4
typing-inspect==0.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.23.2
watchdog==4.0.1
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
yarg==0.1.9
yarl==1.9.4
